+++ packages/api-server/api_server/__main__.py
@@ -9,6 +9,7 @@ def main():
         sio_app,
         host=app_config.host,
         port=app_config.port,
+        log_level=app_config.log_level.lower(),
     )
