+++ packages/api-server/package.json
@@ -7,7 +7,7 @@
     "prepare": "cmake -B build && cmake --build build",
     "test": "python3 -m pipenv run python -m unittest",
     "test:cov": "python3 -m pipenv run python -m coverage run -m unittest && TEST_COVERAGE=1 npm run test:integration",
-    "test:integration": "python3 -m pipenv run python -m unittest discover -s tests",
+    "test:integration": "python3 -m pipenv run python -m unittest discover -s tests -t .",
     "test:report": "python3 -m pipenv run python -m coverage html && xdg-open htmlcov/index.html",
     "lint": "python3 -m pipenv run pylint --ignore=mixins api_server"
   },
